# openapi/init.yml
openapi: 3.0.0
info:
  title: Pikachu API
  version: "1.0.0"
  description: Pikachu API
servers:
  - url: http://localhost:33333/api/v1
    description: Local server
paths:
  /auth/signin:
    post:
      summary: Signin
      description: Signin, return access token, put it in Authorization header
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Signin"
      responses:
        "200":
          description: Custom object created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      resultData:
                        $ref: "#/components/schemas/Token"
        default:
          $ref: "#/components/responses/Response"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 
  responses:
    Response:
      description: Standard API response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Response"
  schemas:
    Response:
      type: object
      properties:
        trid:
          type: string
          description: Transaction ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        resultCode:
          type: string
          description: Result Code
          example: "200"
        resultMsg:
          type: string
          description: Result Message
          example: "result message"
        resultData:
          type: object
          description: Dynamic field for additional data
    Token:
      type: object
      properties:
        accessToken:
          type: string
          description: Access Token
    Signin:
      type: object
      properties:
        email:
          type: string
          description: Email of the user
          example: "test@test.com"
        password:
          type: string
          description: Password of the user
          example: "test123"